updated simulate data
modified 00 to simulate data with the calculated mean value from the actual data, used gamma distribution because my mean value for the lead amount is very small in the unit of ppm, with poison distribution, the simulated value will all be 0. Moreover, this is not a discrete count number, but a continuous measured value, therefore, using poison distribution will only give us a integer number instead of a more accurate simulation with regard to the actual data.